requirements.txt
Django==4.2.2
django-environ==0.10.0
django-storages[google]==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
Pillow==9.5.0
psycopg[binary]==3.1.9
